api/requirements.txt
Django==4.0.6
djangorestframework
djangorestframework-simplejwt
django-cors-headers
psycopg2-binary
firebase-admin
gunicorn
factory_boy
